{"types":"C","runs":11}

+++ Problem35.java
@@ -5,7 +5,7 @@ public class Problem35 extends ConsoleProgram {
 	private static final int NUM_EXPERIMENTS = 10000;
 	private RandomGenerator rgen = RandomGenerator.getInstance();
 	public void run(){
-		int sumFlips = 0;
+		double sumFlips = 0;
 		for(int i = 0; i < NUM_EXPERIMENTS; i++){
 			int currFlips = holdExperiment();
 			sumFlips += currFlips;
